requirements.txt
beautifulsoup4==4.10.0 
kafka-python==2.0.2    
numpy==1.22.3            
pandas==1.4.0            
requests==2.26.0 